+++ packages/paths/src/interpolate-path/array-of-length.ts
@@ -1,0 +1,8 @@
+export function arrayOfLength<T>(length: number, value: T): T[] {
+	const array = Array(length);
+	for (let i = 0; i < length; i++) {
+		array[i] = value;
+	}
+
+	return array;
+}
